Changed the daily report ui to use the html query string argument day_start in its url links.
This is necessary since when you currently get a daily report email, the links
provided therein take you to the LNT daily report page for the proper day but
the offset is set at 16:00. This makes copy pasting the report url into an email
more difficult than it should be.