Fix accesses to uninitialized stack space in alti.isamax.c.
Hal Finkel recently discovered this test could fail when stack pages
are not pre-initialized to zero.  We eventually traced it down to a
loop in the source that accessed values outside the initialized
portion of an array.  With that fixed, we found another problem in the
epilogue of the same loop that caused some elements of the array to be
wrongly ignored.  The test now passes.