New llvm intrinsics to interpret some opcodes.
isSecondaryType is now handled by a function inside the binary. This mean that the Jitter only emit a call to the function instead of the function body.

It is the same in opcode AASTORE: Now, we check if the object is assignable to the aray in a function.
(cherry picked from commit bac3efd9cb3a4c582796da9fdb1187ac41a170ce)